pyproject.toml
[project]
name = "strands-chatbot"
version = "0.1.0"
description = "智能聊天机器人 - 基于 Strands Agent + AWS AgentCore"
requires-python = ">=3.10"
dependencies = [
    "strands-agents>=1.29.0",
    "strands-agents-tools>=0.1.0",
    "boto3>=1.35.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.black]
line-length = 100
target-version = ["py310"]
